+++ src/ext/java/org/opentripplanner/ext/transmodelapi/TransmodelGraphQLSchema.java
@@ -64,6 +64,7 @@ import org.opentripplanner.ext.transmodelapi.model.network.JourneyPatternType;
 import org.opentripplanner.ext.transmodelapi.model.network.LineType;
 import org.opentripplanner.ext.transmodelapi.model.network.PresentationType;
 import org.opentripplanner.ext.transmodelapi.model.network.StopToStopGeometryType;
+import org.opentripplanner.ext.transmodelapi.model.plan.ElevationProfileStepType;
 import org.opentripplanner.ext.transmodelapi.model.plan.LegType;
 import org.opentripplanner.ext.transmodelapi.model.plan.PathGuidanceType;
 import org.opentripplanner.ext.transmodelapi.model.plan.PlanPlaceType;
@@ -318,7 +319,8 @@ public class TransmodelGraphQLSchema {
       rentalVehicleType,
       quayType
     );
-    GraphQLObjectType pathGuidanceType = PathGuidanceType.create();
+    GraphQLObjectType elevationStepType = ElevationProfileStepType.create();
+    GraphQLObjectType pathGuidanceType = PathGuidanceType.create(elevationStepType);
     GraphQLObjectType legType = LegType.create(
       bookingArrangementType,
       interchangeType,
@@ -333,6 +335,7 @@ public class TransmodelGraphQLSchema {
       ptSituationElementType,
       placeType,
       pathGuidanceType,
+      elevationStepType,
       gqlUtil
     );
     GraphQLObjectType tripPatternType = TripPatternType.create(systemNoticeType, legType, gqlUtil);