+++ src/main/java/org/opentripplanner/netex/mapping/TransportModeMapper.java
@@ -6,7 +6,7 @@ import org.rutebanken.netex.model.AllVehicleModesOfTransportEnumeration;
 import org.rutebanken.netex.model.TransportSubmodeStructure;
 
 /**
- * This is a best effort at mapping the NeTEx transport modes to the OTP route codes which are
+ * This is a best-effort at mapping the NeTEx transport modes to the OTP route codes which are
  * identical to the
  * <a href="https://developers.google.com/transit/gtfs/reference/extended-route-types">GTFS extended
  * route types</a>
@@ -31,12 +31,13 @@ class TransportModeMapper {
     }
     return switch (mode) {
       case AIR -> TransitMode.AIRPLANE;
-      case BUS, TAXI -> TransitMode.BUS;
+      case BUS -> TransitMode.BUS;
       case CABLEWAY -> TransitMode.CABLE_CAR;
       case COACH -> TransitMode.COACH;
       case FUNICULAR -> TransitMode.FUNICULAR;
       case METRO -> TransitMode.SUBWAY;
       case RAIL -> TransitMode.RAIL;
+      case TAXI -> TransitMode.TAXI;
       case TRAM -> TransitMode.TRAM;
       case WATER -> TransitMode.FERRY;
       default -> throw new UnsupportedModeException(mode);