+++ src/main/java/org/opentripplanner/routing/core/BicycleOptimizeType.java
@@ -1,5 +1,9 @@
 package org.opentripplanner.routing.core;
 
+import java.util.Collections;
+import java.util.EnumSet;
+import java.util.Set;
+
 /**
  * When planning a bicycle route what should be optimized for. Optimize types are basically
  * combined presets of routing parameters, except for triangle.
@@ -9,5 +13,16 @@ public enum BicycleOptimizeType {
   SAFE,
   FLAT,/* needs a rewrite */
   GREENWAYS,
-  TRIANGLE,
+  TRIANGLE;
+
+  private static final Set<BicycleOptimizeType> NON_TRIANGLE_VALUES = Collections.unmodifiableSet(
+    EnumSet.complementOf(EnumSet.of(TRIANGLE))
+  );
+
+  /**
+   * Return all values that are not {@link BicycleOptimizeType#TRIANGLE}.
+   */
+  public static Set<BicycleOptimizeType> nonTriangleValues() {
+    return NON_TRIANGLE_VALUES;
+  }
 }