+++ docs/Changelog.md
@@ -47,6 +47,11 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add GTFS stop sequence to GTFS GraphQL API [#5153](https://github.com/opentripplanner/OpenTripPlanner/pull/5153)
 - Remove walk leg in a stay seated transfer [#5135](https://github.com/opentripplanner/OpenTripPlanner/pull/5135)
 - Validate distinct from/to temporary vertices [#5181](https://github.com/opentripplanner/OpenTripPlanner/pull/5181)
+- Add support for taxi mode [#5183](https://github.com/opentripplanner/OpenTripPlanner/pull/5183)
+- Fix bike triangle in Transmodel API [#5179](https://github.com/opentripplanner/OpenTripPlanner/pull/5179)
+- Bug fixes in stop area relation processing [#5166](https://github.com/opentripplanner/OpenTripPlanner/pull/5166)
+- Enable GTFS GraphQL API by default, remove the word  from its documentation [#5202](https://github.com/opentripplanner/OpenTripPlanner/pull/5202)
+- Allow underscores in GTFS feed IDs [#5191](https://github.com/opentripplanner/OpenTripPlanner/pull/5191)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)